Add simple search functionality to DBHelper and Actionbar. NO RESULTS DISPLAYED YET.
I'm just trying to come up with a way to query the DB and get a list of articles for a search term. That's an important usecase to me, and I think it'd be valuable to other users too.

My problem at the moment is that I don't know how to display the list of articles. I need to replace one of the currently displayed fragments, but at the moment I can only come up with my own ListView created in-memory. I'm limiting the search results to 100 at the moment, so it's not that much of an issue, but it's a pretty artificial limit and I want to get around by using a real Cursor.