Download feed icons and display them in feed list. Fixes #241
URL path to feed icons can be set via Preferences, getting it from
the server config is not implemented. This can be improved, but I was
not clear about when to update the value and I didn't want to fetch it
for every request to the server or something similar naive.

The feed icon is displayed in the 2nd line of the feed item list, on
the right of the item date. The layout is not nice and centered, but I
had to give up on the Android layout...